RSpec files corrected
